Removed internal 'good' system
With winAgainst and lostTo recorded the Option Data doesn't really need to know about what was "good" or a default value. That lets the default be handled at the higher level of the ranking.

Makes power ranking difficult to display internally, so need to add display classes to manage sharing that info.